+++ packages/manager/src/featureFlags.ts
@@ -92,7 +92,6 @@ export interface Flags {
   databases: boolean;
   dbaasV2: BetaFeatureFlag;
   disableLargestGbPlans: boolean;
-  eventMessagesV2: boolean;
   gecko: boolean; // @TODO gecko: delete this after next release
   gecko2: GaFeatureFlag;
   gpuv2: gpuV2;
