+++ packages/manager/src/features/Events/EventsLanding.tsx
@@ -11,11 +11,9 @@ import { TableRowEmpty } from 'src/components/TableRowEmpty/TableRowEmpty';
 import { TableRowError } from 'src/components/TableRowError/TableRowError';
 import { TableRowLoading } from 'src/components/TableRowLoading/TableRowLoading';
 import { EVENTS_LIST_FILTER } from 'src/features/Events/constants';
-import { useFlags } from 'src/hooks/useFlags';
 import { useEventsInfiniteQuery } from 'src/queries/events/events';
 
 import { EventRow } from './EventRow';
-import { EventRowV2 } from './EventRowV2';
 import {
   StyledH1Header,
   StyledLabelTableCell,
@@ -32,7 +30,6 @@ interface Props {
 
 export const EventsLanding = (props: Props) => {
   const { emptyMessage, entityId } = props;
-  const flags = useFlags();
 
   const filter: Filter = { ...EVENTS_LIST_FILTER };
 
@@ -71,21 +68,13 @@ export const EventsLanding = (props: Props) => {
     } else {
       return (
         <>
-          {events?.map((event) =>
-            flags.eventMessagesV2 ? (
-              <EventRowV2
-                entityId={entityId}
-                event={event}
-                key={`event-${event.id}`}
-              />
-            ) : (
-              <EventRow
-                entityId={entityId}
-                event={event}
-                key={`event-${event.id}`}
-              />
-            )
-          )}
+          {events?.map((event) => (
+            <EventRow
+              entityId={entityId}
+              event={event}
+              key={`event-${event.id}`}
+            />
+          ))}
           {isFetchingNextPage && (
             <TableRowLoading
               columns={4}
@@ -105,19 +94,12 @@ export const EventsLanding = (props: Props) => {
       <Table aria-label="List of Events">
         <TableHead>
           <TableRow>
-            {!flags.eventMessagesV2 && (
-              <Hidden smDown>
-                <TableCell style={{ padding: 0, width: '1%' }} />
-              </Hidden>
-            )}
             <StyledLabelTableCell>Event</StyledLabelTableCell>
-            {flags.eventMessagesV2 && (
-              <Hidden smDown>
-                <TableCell data-qa-events-username-header sx={{ width: 150 }}>
-                  User
-                </TableCell>
-              </Hidden>
-            )}
+            <Hidden smDown>
+              <TableCell data-qa-events-username-header sx={{ width: 150 }}>
+                User
+              </TableCell>
+            </Hidden>
             <StyledTableCell sx={{ width: 175 }}>Relative Date</StyledTableCell>
             <Hidden mdDown>
               <StyledTableCell data-qa-events-time-header sx={{ width: 150 }}>
