+++ packages/manager/src/features/NotificationCenter/NotificationData/RenderEvent.tsx
@@ -1,20 +1,21 @@
-// TODO eventMessagesV2: delete when flag is removed
-import { Event } from '@linode/api-v4/lib/account/types';
 import * as React from 'react';
 
+import { BarPercent } from 'src/components/BarPercent';
 import { Box } from 'src/components/Box';
-import { Divider } from 'src/components/Divider';
-import { HighlightedMarkdown } from 'src/components/HighlightedMarkdown/HighlightedMarkdown';
 import { Typography } from 'src/components/Typography';
-import { getEventTimestamp } from 'src/utilities/eventUtils';
-import { getAllowedHTMLTags } from 'src/utilities/sanitizeHTML.utils';
+import {
+  formatProgressEvent,
+  getEventMessage,
+  getEventUsername,
+} from 'src/features/Events/utils';
 
 import {
-  RenderEventGravatar,
+  RenderEventGravatarV2,
   RenderEventStyledBox,
   useRenderEventStyles,
 } from './RenderEvent.styles';
-import useEventInfo from './useEventInfo';
+
+import type { Event } from '@linode/api-v4/lib/account/types';
 
 interface RenderEventProps {
   event: Event;
@@ -22,40 +23,44 @@ interface RenderEventProps {
 }
 
 export const RenderEvent = React.memo((props: RenderEventProps) => {
-  const { classes, cx } = useRenderEventStyles();
   const { event } = props;
-  const { message } = useEventInfo(event);
+  const { classes, cx } = useRenderEventStyles();
+  const unseenEventClass = cx({ [classes.unseenEventV2]: !event.seen });
+  const message = getEventMessage(event);
+  const username = getEventUsername(event);
 
-  const unseenEventClass = cx({ [classes.unseenEvent]: !event.seen });
-
+  /**
+   * Some event types may not be handled by our system (or new types or new ones may be added that we haven't caught yet).
+   * Filter these out so we don't display blank messages to the user.
+   * We have Sentry events being logged for these cases, so we can always go back and add support for them as soon as we become aware.
+   */
   if (message === null) {
     return null;
   }
 
-  const eventMessage = (
-    <div className={unseenEventClass}>
-      <HighlightedMarkdown
-        sanitizeOptions={{
-          ALLOWED_TAGS: getAllowedHTMLTags('strict'),
-          disallowedTagsMode: 'discard',
-        }}
-        textOrMarkdown={message}
-      />
-    </div>
-  );
+  const { progressEventDisplay, showProgress } = formatProgressEvent(event);
 
   return (
-    <>
-      <RenderEventStyledBox data-test-id={event.action} display="flex">
-        <RenderEventGravatar username={event.username} />
-        <Box sx={{ marginTop: '-2px' }}>
-          {eventMessage}
-          <Typography className={unseenEventClass}>
-            {getEventTimestamp(event).toRelative()}
-          </Typography>
-        </Box>
-      </RenderEventStyledBox>
-      <Divider />
-    </>
+    <RenderEventStyledBox
+      className={unseenEventClass}
+      data-testid={event.action}
+    >
+      <RenderEventGravatarV2 username={event.username} />
+      <Box sx={{ marginTop: '-2px', paddingRight: 1, width: '100%' }}>
+        {message}
+        {showProgress && (
+          <BarPercent
+            className={classes.bar}
+            max={100}
+            narrow
+            rounded
+            value={event.percent_complete ?? 0}
+          />
+        )}
+        <Typography sx={{ fontSize: '0.8rem' }}>
+          {progressEventDisplay} | {username}
+        </Typography>
+      </Box>
+    </RenderEventStyledBox>
   );
 });
