--- packages/manager/src/features/NotificationCenter/index.tsx
@@ -1,1 +1,0 @@
-export { default as Events } from './Events';
