+++ packages/manager/src/features/TopMenu/NotificationMenu/NotificationMenu.test.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 
 import { renderWithTheme } from 'src/utilities/testHelpers';
 
-import { NotificationMenuV2 } from './NotificationMenuV2';
+import { NotificationMenu } from './NotificationMenu';
 
 describe('NotificationMenuV2', () => {
   // Very basic unit - the functionality is tested in the integration test
   it('should render', () => {
-    const { getByRole } = renderWithTheme(<NotificationMenuV2 />);
+    const { getByRole } = renderWithTheme(<NotificationMenu />);
 
     expect(getByRole('button', { name: 'Notifications' })).toBeInTheDocument();
   });
