+++ packages/manager/src/features/TopMenu/NotificationMenu/NotificationMenu.tsx
@@ -1,58 +1,55 @@
-// TODO eventMessagesV2: delete when flag is removed
+import AutorenewIcon from '@mui/icons-material/Autorenew';
 import { IconButton } from '@mui/material';
 import Popover from '@mui/material/Popover';
 import { styled } from '@mui/material/styles';
 import * as React from 'react';
-import { useDispatch } from 'react-redux';
+import { useHistory } from 'react-router-dom';
 
 import Bell from 'src/assets/icons/notification.svg';
+import { Box } from 'src/components/Box';
 import { Chip } from 'src/components/Chip';
-import Events from 'src/features/NotificationCenter/Events';
+import { Divider } from 'src/components/Divider';
+import { LinkButton } from 'src/components/LinkButton';
+import { Typography } from 'src/components/Typography';
 import {
   notificationContext as _notificationContext,
   menuButtonId,
 } from 'src/features/NotificationCenter/NotificationContext';
-import { useEventNotifications } from 'src/features/NotificationCenter/NotificationData/useEventNotifications';
+import { RenderEvent } from 'src/features/NotificationCenter/NotificationData/RenderEvent';
 import { useFormattedNotifications } from 'src/features/NotificationCenter/NotificationData/useFormattedNotifications';
 import Notifications from 'src/features/NotificationCenter/Notifications';
 import { useDismissibleNotifications } from 'src/hooks/useDismissibleNotifications';
 import { usePrevious } from 'src/hooks/usePrevious';
 import { useNotificationsQuery } from 'src/queries/account/notifications';
-import { useMarkEventsAsSeen } from 'src/queries/events/events';
-import { ThunkDispatch } from 'src/store/types';
+import { isInProgressEvent } from 'src/queries/events/event.helpers';
+import {
+  useEventsInfiniteQuery,
+  useMarkEventsAsSeen,
+} from 'src/queries/events/events';
+import { rotate360 } from 'src/styles/keyframes';
 
 import { TopMenuTooltip, topMenuIconButtonSx } from '../TopMenuTooltip';
 
-const StyledChip = styled(Chip)(() => ({
-  '& .MuiChip-label': {
-    paddingLeft: 2,
-    paddingRight: 2,
-  },
-  fontSize: '0.72rem',
-  height: '1rem',
-  justifyContent: 'center',
-  left: 20,
-  padding: 0,
-  position: 'absolute',
-  top: 4,
-}));
-
 export const NotificationMenu = () => {
+  const history = useHistory();
   const { dismissNotifications } = useDismissibleNotifications();
   const { data: notifications } = useNotificationsQuery();
   const formattedNotifications = useFormattedNotifications();
-  const eventNotifications = useEventNotifications();
   const notificationContext = React.useContext(_notificationContext);
+
+  const { data, events } = useEventsInfiniteQuery();
   const { mutateAsync: markEventsAsSeen } = useMarkEventsAsSeen();
 
   const numNotifications =
-    eventNotifications.filter((thisEvent) => thisEvent.countInTotal).length +
-    formattedNotifications.filter((thisEvent) => thisEvent.countInTotal).length;
+    (events?.filter((event) => !event.seen).length ?? 0) +
+    formattedNotifications.filter(
+      (notificationItem) => notificationItem.countInTotal
+    ).length;
+
+  const showInProgressEventIcon = events?.some(isInProgressEvent);
 
   const anchorRef = React.useRef<HTMLButtonElement>(null);
   const prevOpen = usePrevious(notificationContext.menuOpen);
-
-  const dispatch = useDispatch<ThunkDispatch>();
 
   const handleNotificationMenuToggle = () => {
     if (!notificationContext.menuOpen) {
@@ -69,20 +66,12 @@ export const NotificationMenu = () => {
   React.useEffect(() => {
     if (prevOpen && !notificationContext.menuOpen) {
       // Dismiss seen notifications after the menu has closed.
-      if (eventNotifications.length > 0) {
-        markEventsAsSeen(eventNotifications[0].eventId);
+      if (events && events.length >= 1 && !events[0].seen) {
+        markEventsAsSeen(events[0].id);
       }
       dismissNotifications(notifications ?? [], { prefix: 'notificationMenu' });
     }
-  }, [
-    notificationContext.menuOpen,
-    dismissNotifications,
-    eventNotifications,
-    notifications,
-    dispatch,
-    prevOpen,
-    markEventsAsSeen,
-  ]);
+  }, [notificationContext.menuOpen]);
 
   const id = notificationContext.menuOpen ? 'notifications-popover' : undefined;
 
@@ -103,7 +92,16 @@ export const NotificationMenu = () => {
         >
           <Bell height="20px" width="20px" />
           {numNotifications > 0 && (
-            <StyledChip color="success" label={numNotifications} size="small" />
+            <StyledChip
+              color="primary"
+              data-testid="events-count-notification"
+              label={numNotifications > 9 ? '9+' : numNotifications}
+              showPlus={numNotifications > 9}
+              size="small"
+            />
+          )}
+          {showInProgressEventIcon && (
+            <StyledAutorenewIcon data-testid="in-progress-event-icon" />
           )}
         </IconButton>
       </TopMenuTooltip>
@@ -132,8 +130,53 @@ export const NotificationMenu = () => {
         open={notificationContext.menuOpen}
       >
         <Notifications />
-        <Events />
+        <Box>
+          <Box display="flex" justifyContent="space-between" px={2}>
+            <Typography variant="h3">Events</Typography>
+            <LinkButton
+              onClick={() => {
+                history.push('/events');
+                handleClose();
+              }}
+            >
+              View all events
+            </LinkButton>
+          </Box>
+          <Divider spacingBottom={0} />
+          {data?.pages[0].data.slice(0, 20).map((event) => (
+            <RenderEvent event={event} key={event.id} onClose={handleClose} />
+          ))}
+        </Box>
       </Popover>
     </>
   );
 };
+
+const StyledChip = styled(Chip, {
+  label: 'StyledEventNotificationChip',
+  shouldForwardProp: (prop) => prop !== 'showPlus',
+})<{ showPlus: boolean }>(({ theme, ...props }) => ({
+  '& .MuiChip-label': {
+    paddingLeft: 2,
+    paddingRight: 2,
+  },
+  borderRadius: props.showPlus ? 12 : '50%',
+  fontFamily: theme.font.bold,
+  fontSize: '0.72rem',
+  height: 18,
+  justifyContent: 'center',
+  left: 20,
+  padding: 0,
+  position: 'absolute',
+  top: 0,
+  width: props.showPlus ? 22 : 18,
+}));
+
+const StyledAutorenewIcon = styled(AutorenewIcon)(({ theme }) => ({
+  animation: `${rotate360} 2s linear infinite`,
+  bottom: 4,
+  color: theme.palette.primary.main,
+  fontSize: 18,
+  position: 'absolute',
+  right: 2,
+}));
